Refactor: Seperate Header, Modal from Pureplate
